Added Colors.cs file
