+++ TelCo.ColorCoder/Utils/Colors.cs
@@ -1,0 +1,126 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Diagnostics;
+using System.Drawing;
+using System.Threading;
+
+namespace TelCo.ColorCoder.Utils
+{
+    class Colors
+    {
+        /// <summary>
+        /// Array of Major colors
+        /// </summary>
+        private static Color[] colorMapMajor;
+        /// <summary>
+        /// Array of minor colors
+        /// </summary>
+        private static Color[] colorMapMinor;
+        /// <summary>
+        /// data type defined to hold the two colors of clor pair
+        /// </summary>
+        internal class ColorPair
+        {
+            internal Color majorColor;
+            internal Color minorColor;
+            public override string ToString()
+            {
+                return string.Format("MajorColor:{0}, MinorColor:{1}", majorColor.Name, minorColor.Name);
+            }
+        }
+        /// <summary>
+        /// Static constructor required to initialize static variable
+        /// </summary>
+        static Colors()
+        {
+            colorMapMajor = new Color[] { Color.White, Color.Red, Color.Black, Color.Yellow, Color.Violet };
+            colorMapMinor = new Color[] { Color.Blue, Color.Orange, Color.Green, Color.Brown, Color.SlateGray };
+        }
+
+        /// <summary>
+        /// Given a pair number function returns the major and minor colors in that order
+        /// </summary>
+        /// <param name="pairNumber">Pair number of the color to be fetched</param>
+        /// <returns></returns>
+        public static ColorPair GetColorFromPairNumber(int pairNumber)
+        {
+            // The function supports only 1 based index. Pair numbers valid are from 1 to 25
+            int minorSize = colorMapMinor.Length;
+            int majorSize = colorMapMajor.Length;
+            if (pairNumber < 1 || pairNumber > minorSize * majorSize)
+            {
+                throw new ArgumentOutOfRangeException(
+                    string.Format("Argument PairNumber:{0} is outside the allowed range", pairNumber));
+            }
+
+            // Find index of major and minor color from pair number
+            int zeroBasedPairNumber = pairNumber - 1;
+            int majorIndex = zeroBasedPairNumber / minorSize;
+            int minorIndex = zeroBasedPairNumber % minorSize;
+
+            // Construct the return val from the arrays
+            ColorPair pair = new ColorPair()
+            {
+                majorColor = colorMapMajor[majorIndex],
+                minorColor = colorMapMinor[minorIndex]
+            };
+
+            // return the value
+            return pair;
+        }
+        /// <summary>
+        /// Given the two colors the function returns the pair number corresponding to them
+        /// </summary>
+        /// <param name="pair">Color pair with major and minor color</param>
+        /// <returns></returns>
+        public static int GetPairNumberFromColor(ColorPair pair)
+        {
+            // Find the major color in the array and get the index
+            int majorIndex = -1;
+            for (int i = 0; i < colorMapMajor.Length; i++)
+            {
+                if (colorMapMajor[i] == pair.majorColor)
+                {
+                    majorIndex = i;
+                    break;
+                }
+            }
+
+            // Find the minor color in the array and get the index
+            int minorIndex = -1;
+            for (int i = 0; i < colorMapMinor.Length; i++)
+            {
+                if (colorMapMinor[i] == pair.minorColor)
+                {
+                    minorIndex = i;
+                    break;
+                }
+            }
+            // If colors can not be found throw an exception
+            if (majorIndex == -1 || minorIndex == -1)
+            {
+                throw new ArgumentException(
+                    string.Format("Unknown Colors: {0}", pair.ToString()));
+            }
+
+            // Compute pair number and Return  
+            // (Note: +1 in compute is because pair number is 1 based, not zero)
+            return (majorIndex * colorMapMinor.Length) + (minorIndex + 1);
+        }
+        public override string ToString() 
+        {
+            string colorCode = "";
+            int i = 1;
+            foreach(var majorColor in colorMapMajor)
+            {
+                foreach(var minorColor in colorMapMinor)
+                {
+                    colorCode +=$"MajorColor:{majorColor}, MinorColor:{minorColor}, PairNumber:{i++}\n";
+                }
+            }
+            return colorCode;
+            
+        }
+    }
+}
